Add support for overlayfs and cleanup code for future storage drivers
This cleans up the code a bit so that storage driver specific details
and actions are self-contained in their own classes.

It adds support for --overlay to use the overlayfs driver. A quirk of
this is that selinux isn't supported inside the container when using
overlayfs. The OPTION flag will be edited to remove --selinux-enabled if
it is present.